Fix update

+++ app/db/requests.py
@@ -2,7 +2,7 @@ from datetime import datetime, timedelta
 from zoneinfo import ZoneInfo
 import uuid
 
-from sqlalchemy import select, BigInteger, func
+from sqlalchemy import select, int, func
 
 from app.db.engine import async_session
 from app.db.models import (
@@ -22,7 +22,7 @@ from app.db.models import (
 
 # ----------------- LOT FUNCTIONS -----------------
 
-async def set_lot(tid: BigInteger, starter_price: int, hours_exp: int, pid: str, blitz_price: int):
+async def set_lot(tid: int, starter_price: int, hours_exp: int, pid: str, blitz_price: int):
     async with async_session() as session:
         user = await session.scalar(
             select(UserBase)
@@ -49,7 +49,7 @@ async def get_lot_by_uuid(uuid: str):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.uuid == uuid))
 
-async def get_lot_data(lot_id: BigInteger):
+async def get_lot_data(lot_id: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.id == lot_id))
 
@@ -57,14 +57,14 @@ async def get_lot_data_by_photo_id(pid: str):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.photo_id == pid))
 
-async def approve_lot(lot_id: BigInteger):
+async def approve_lot(lot_id: int):
     async with async_session() as session:
         lot = await session.scalar(select(LotBase).where(LotBase.id == lot_id).with_for_update())
         lot.is_post = LotModStatus.APPROVED
         lot.expired_at = datetime.now(ZoneInfo("Europe/Moscow")).replace(tzinfo=None) + timedelta(hours=lot.expired_time)
         await session.commit()
 
-async def reject_lot(lot_id: BigInteger):
+async def reject_lot(lot_id: int):
     async with async_session() as session:
         lot = await session.scalar(select(LotBase).where(LotBase.id == lot_id).with_for_update())
         lot.is_post = LotModStatus.REJECTED
@@ -82,27 +82,27 @@ async def get_previous_lot(current_lot_id):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.is_post == LotModStatus.PENDING, LotBase.id < current_lot_id).order_by(LotBase.id.desc()).limit(1))
 
-async def get_first_user_lot(tid: BigInteger):
+async def get_first_user_lot(tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.seller == tid).order_by(LotBase.id.asc()).limit(1))
 
-async def get_next_user_lot(current_lot_id, tid: BigInteger):
+async def get_next_user_lot(current_lot_id, tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.seller == tid, LotBase.id > current_lot_id).order_by(LotBase.id.asc()).limit(1))
 
-async def get_previous_user_lot(current_lot_id, tid: BigInteger):
+async def get_previous_user_lot(current_lot_id, tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.seller == tid, LotBase.id < current_lot_id).order_by(LotBase.id.desc()).limit(1))
 
-async def get_first_user_lot_bid(tid: BigInteger):
+async def get_first_user_lot_bid(tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.applicant == tid, LotBase.status == LotStatus.TRADING).order_by(LotBase.id.asc()).limit(1))
 
-async def get_next_user_lot_bid(current_lot_id, tid: BigInteger):
+async def get_next_user_lot_bid(current_lot_id, tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.applicant == tid, LotBase.id > current_lot_id, LotBase.status == LotStatus.TRADING).order_by(LotBase.id.asc()).limit(1))
 
-async def get_previous_user_lot_bid(current_lot_id, tid: BigInteger):
+async def get_previous_user_lot_bid(current_lot_id, tid: int):
     async with async_session() as session:
         return await session.scalar(select(LotBase).where(LotBase.applicant == tid, LotBase.id < current_lot_id, LotBase.status == LotStatus.TRADING).order_by(LotBase.id.desc()).limit(1))
 
@@ -140,7 +140,7 @@ async def buy_now(lot_id, user_id, real_price: int):
 
 # ----------------- USER FUNCTIONS -----------------
 
-async def get_user_data(tg_id: BigInteger):
+async def get_user_data(tg_id: int):
     async with async_session() as session:
         return await session.scalar(select(UserBase).where(UserBase.telegram_id == tg_id))
 
@@ -151,7 +151,7 @@ async def get_user_data_ref(ref_code: str):
             print("Referral is not found")
         return await session.scalar(select(UserBase).where(UserBase.id == referral.user_id))
 
-async def get_user_data_id(id: BigInteger):
+async def get_user_data_id(id: int):
     async with async_session() as session:
         return await session.scalar(select(UserBase).where(UserBase.id == id))
 
@@ -159,19 +159,19 @@ async def get_user_by_username(username: str):
     async with async_session() as session:
         return await session.scalar(select(UserBase).where(UserBase.username == username))
 
-async def set_new_user(tid: BigInteger):
+async def set_new_user(tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid))
         user.is_new = False
         await session.commit()
 
-async def deposit_balance(tg_id: BigInteger, stars: int):
+async def deposit_balance(tg_id: int, stars: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tg_id).with_for_update())
         user.balance += stars
         await session.commit()
 
-async def increase_balance(tg_id: BigInteger, bid: int):
+async def increase_balance(tg_id: int, bid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tg_id).with_for_update())
         if bid <=2000000:
@@ -180,7 +180,7 @@ async def increase_balance(tg_id: BigInteger, bid: int):
             print("Value of bid too high")
         await session.commit()
 
-async def decrease_balance(tg_id: BigInteger, bid: int):
+async def decrease_balance(tg_id: int, bid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tg_id).with_for_update())
         if bid <= 2000000:
@@ -189,26 +189,26 @@ async def decrease_balance(tg_id: BigInteger, bid: int):
             print("Value of bid too high")
         await session.commit()
 
-async def ban_user(tid: BigInteger):
+async def ban_user(tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         user.is_banned = True
         await session.commit()
 
-async def unban_user(tid: BigInteger):
+async def unban_user(tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         user.is_banned = False
         await session.commit()
 
-async def warn_user(utid: BigInteger, atid: BigInteger, reason: str):
+async def warn_user(utid: int, atid: int, reason: str):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == utid).with_for_update())
         admin = await session.scalar(select(UserBase).where(UserBase.telegram_id == atid).with_for_update())
         session.add(WarnBase(user_id = user.id, admin_id = admin.id, reason=reason))
         await session.commit()
 
-async def warn_count(tid: BigInteger):
+async def warn_count(tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         count_result = await session.execute(
@@ -247,13 +247,13 @@ async def get_blocked_users():
 
 # ----------------- REFERRAL FUNCTIONS -----------------
 
-async def add_new_referral_link(referral_link: str, tid: BigInteger):
+async def add_new_referral_link(referral_link: str, tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         session.add(ReferralBase(user_id=user.id, link=referral_link))
         await session.commit()
 
-async def set_user_referral(referral_id: BigInteger, tid: BigInteger):
+async def set_user_referral(referral_id: int, tid: int):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         user.ref_id = referral_id
@@ -270,7 +270,7 @@ async def get_blank_data(bid: int):
     async with async_session() as session:
         return await session.scalar(select(WithdrawalRequest).where(WithdrawalRequest.id == bid))
 
-async def add_new_blank(tid: BigInteger ,stars: int, bank: str, number: str):
+async def add_new_blank(tid: int ,stars: int, bank: str, number: str):
     async with async_session() as session:
         user = await session.scalar(select(UserBase).where(UserBase.telegram_id == tid).with_for_update())
         if bank == "tinkoff":
@@ -315,7 +315,7 @@ async def get_previous_blank(current_blank_id):
     async with async_session() as session:
         return await session.scalar(select(WithdrawalRequest).where(WithdrawalRequest.status == BlankModStatus.PENDING, WithdrawalRequest.id < current_blank_id).order_by(WithdrawalRequest.id.desc()).limit(1))
 
-async def approve_blank(photo_id: str, admin_id: BigInteger, blank_id: int):
+async def approve_blank(photo_id: str, admin_id: int, blank_id: int):
     async with async_session() as session:
         blank = await session.scalar(select(WithdrawalRequest).where(WithdrawalRequest.id == blank_id))
         admin = await session.scalar(select(UserBase).where(UserBase.telegram_id == admin_id))
@@ -325,7 +325,7 @@ async def approve_blank(photo_id: str, admin_id: BigInteger, blank_id: int):
         blank.status = BlankModStatus.APPROVED
         await session.commit()
 
-async def reject_blank(admin_id: BigInteger, blank_id: int):
+async def reject_blank(admin_id: int, blank_id: int):
     async with async_session() as session:
         blank = await session.scalar(select(WithdrawalRequest).where(WithdrawalRequest.id == blank_id))
         admin = await session.scalar(select(UserBase).where(UserBase.telegram_id == admin_id))
@@ -368,7 +368,7 @@ async def get_previous_dispute(current_dispute_id):
     async with async_session() as session:
         return await session.scalar(select(DisputeBase).where(DisputeBase.status == DisputeStatusEnum.PENDING, DisputeBase.id < current_dispute_id).order_by(DisputeBase.id.desc()).limit(1))
 
-async def reject_dispute(did: int, aid: BigInteger):
+async def reject_dispute(did: int, aid: int):
     async with async_session() as session:
         admin = await session.scalar(select(UserBase).where(UserBase.telegram_id == aid))
         dispute = await session.scalar(select(DisputeBase).where(DisputeBase.id == did))
@@ -378,7 +378,7 @@ async def reject_dispute(did: int, aid: BigInteger):
         dispute.admin_id = admin.id
         await session.commit()
 
-async def approve_dispute(did: int, aid: BigInteger):
+async def approve_dispute(did: int, aid: int):
     async with async_session() as session:
         admin = await session.scalar(select(UserBase).where(UserBase.telegram_id == aid))
         dispute = await session.scalar(select(DisputeBase).where(DisputeBase.id == did))
